Refactoring unit tests and solving python3 issues.
Now, the unit tests don't mock any data to run the tests.
Actually, the unit tests expects that the certutil command will
be available. This way, the tests will "run the real code" with
real data.
Inside the mock_data directory, there is a couple of certificates
that are used in the unit tests.